fix: Do not rename the migration, fix it and adds a second instance with a new name
This is safer to make sure the new migration runs.
